% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/age.R
\name{age}
\alias{age}
\title{Generate Random Vector of Ages}
\usage{
age(n, x = 20:35, prob = NULL, name = "Age")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random factor vector of gender elements.
}
\description{
Generate a random vector of ages.
}
\examples{
age(10)
hist(age(10000))
interval(age, 3, n = 1000)
}
\seealso{
Other variable.functions: \code{\link{animal}},
  \code{\link{pet}}; \code{\link{answer}};
  \code{\link{area}}; \code{\link{coin}};
  \code{\link{color}}, \code{\link{primary}};
  \code{\link{death}}, \code{\link{died}};
  \code{\link{dice}}; \code{\link{dna}};
  \code{\link{dummy}}; \code{\link{education}};
  \code{\link{gpa}}, \code{\link{grade}},
  \code{\link{grade_letter}}; \code{\link{grade_level}};
  \code{\link{group}}; \code{\link{height}},
  \code{\link{height_cm}}, \code{\link{height_in}};
  \code{\link{income}}; \code{\link{iq}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{lower}},
  \code{\link{lower_factor}}, \code{\link{upper}},
  \code{\link{upper_factor}}; \code{\link{marital}};
  \code{\link{month}}; \code{\link{name}};
  \code{\link{normal}}, \code{\link{normal_round}};
  \code{\link{political}}; \code{\link{race}};
  \code{\link{sat}}; \code{\link{sentence}};
  \code{\link{sex}}; \code{\link{string}};
  \code{\link{valid}}
}
\keyword{age}

