% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{animal_list}
\alias{animal_list}
\title{Animal List}
\format{
A character vector with 591 elements
}
\usage{
data(animal_list)
}
\description{
A dataset containing a character vector animals
}
\references{
https://a-z-animals.com/animals
}
\keyword{datasets}
