% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/group.R
\name{group}
\alias{group}
\title{Generate Random Vector of Groups}
\usage{
group(n, x = c("Control", "Treatment"), prob = NULL, name = "Group")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of groups to sample from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random factor vector of group (control/treatment) elements.
}
\description{
Generate a random vector of groups (control/treatment).
}
\examples{
group(10)
100*table(group(n <- 10000))/n
100*table(group(n <- 10000, prob = c(.3, .7)))/n
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{answer}}; \code{\link{area}};
  \code{\link{coin}}; \code{\link{death}},
  \code{\link{died}}; \code{\link{dice}};
  \code{\link{dna}}; \code{\link{dummy}};
  \code{\link{gpa}}, \code{\link{grade}},
  \code{\link{grade_letter}}; \code{\link{grade_level}};
  \code{\link{height}}, \code{\link{height_cm}},
  \code{\link{height_in}}; \code{\link{income}};
  \code{\link{iq}}; \code{\link{likert}},
  \code{\link{likert_5}}, \code{\link{likert_7}};
  \code{\link{lorem_ipsum}}, \code{\link{paragraph}};
  \code{\link{lower}}, \code{\link{lower_factor}},
  \code{\link{upper}}, \code{\link{upper_factor}};
  \code{\link{marital}}; \code{\link{normal}},
  \code{\link{normal_round}}; \code{\link{political}};
  \code{\link{race}}; \code{\link{sat}}; \code{\link{sex}};
  \code{\link{string}}; \code{\link{valid}}
}
\keyword{group}

