% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/height.R
\name{height}
\alias{height}
\alias{height_cm}
\alias{height_in}
\title{Generate Random Vector of Heights}
\usage{
height(n, mean = 69, sd = 3.75, digits = 0, name = "Height")

height_in(n, mean = 69, sd = 3.75, digits = 1, name = "Height(in)")

height_cm(n, mean = 175.26, sd = 9.525, digits = 1, name = "Height(cm)")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{mean}{The mean value for the normal distribution to be drawn from.}

\item{sd}{The standard deviation of the normal distribution to draw from.}

\item{digits}{Integer indicating the number of decimal places to be used.
Negative values are allowed (see \code{\link[base]{round}}).}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random normal vector of height elements.
}
\description{
\code{height} and \code{height_in} - Generate a random normal vector of
heights in inches.

\code{height_cm} - Generate a random normal vector of heights in centimeters.
}
\note{
\code{height} rounds to nearest whoe number.  \code{height_in} &
\code{height_in} round to the nearest tenths.
}
\examples{
height(10)
hist(height(10000))
interval(height, 5, n = 1000)
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{coin}}; \code{\link{death}},
  \code{\link{died}}; \code{\link{dice}};
  \code{\link{dummy}}; \code{\link{gpa}},
  \code{\link{grade}}, \code{\link{grade_letter}};
  \code{\link{group}}; \code{\link{iq}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{normal}},
  \code{\link{normal_round}}; \code{\link{race}};
  \code{\link{sex}}; \code{\link{string}};
  \code{\link{valid}}
}
\keyword{height}

