% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/iq.R
\name{iq}
\alias{iq}
\title{Generate Random Vector of Intelligence Quotients (IQs)}
\usage{
iq(n, mean = 100, sd = 10, digits = 0, name = "IQ")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{mean}{The mean value for the normal distribution to be drawn from.}

\item{sd}{The standard deviation of the normal distribution to draw from.}

\item{digits}{Integer indicating the number of decimal places to be used.
Negative values are allowed (see \code{\link[base]{round}}).}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random normal vector of IQ elements.
}
\description{
\code{grade} - Generate a random normal vector of intelligence quotients
(IQs).
}
\examples{
iq(10)
hist(iq(10000))
interval(iq, 5, n = 1000)
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{coin}}; \code{\link{death}},
  \code{\link{died}}; \code{\link{dice}};
  \code{\link{dummy}}; \code{\link{gpa}},
  \code{\link{grade}}, \code{\link{grade_letter}};
  \code{\link{group}}; \code{\link{height}},
  \code{\link{height_cm}}, \code{\link{height_in}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{normal}},
  \code{\link{normal_round}}; \code{\link{race}};
  \code{\link{sex}}; \code{\link{string}};
  \code{\link{valid}}
}
\keyword{intelligence}
\keyword{iq}

