% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{languages}
\alias{languages}
\title{Languages of the World}
\format{
A data frame with 99 rows and 4 variables
}
\usage{
data(languages)
}
\description{
A dataset containing native language use statistics taken from:
https://en.wikipedia.org/wiki/List_of_languages_by_number_of_native_speakers.
}
\details{
\itemize{
  \item Language. The language spoken.
  \item N. The number of speakers world-wide.
  \item Proportion. The proportion of speakers.
  \item Percent. The percentage of speakers.
}
}
\references{
https://en.wikipedia.org/wiki/List_of_languages_by_number_of_native_speakers
}
\keyword{datasets}
