% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/likert.R
\name{likert}
\alias{likert}
\alias{likert_5}
\alias{likert_7}
\title{Generate Random Vector of Likert-Type Responses}
\usage{
likert(n, x = c("Strongly Agree ", "Agree", "Neutral", "Disagree",
  "Strongly Disagree"), prob = NULL, name = "Likert")

likert_5(n, x = c("Strongly Agree ", "Agree", "Neutral", "Disagree",
  "Strongly Disagree"), prob = NULL, name = "Likert")

likert_7(n, x = c("Strongly Agree ", "Agree", "Somewhat Agree", "Neutral",
  "Somewhat Disagree", "Disagree", "Strongly Disagree"), prob = NULL,
  name = "Likert")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random vector of Likert-type response elements.
}
\description{
Generate a random vector of Likert-type responses.
}
\note{
\code{likert} & \code{likert_5} are identical outputs, sampeling from a
5-point response scale.  \code{likert_7} samples from a 7-point response
scale.
}
\examples{
dice(10)
barplot(table(dice(10000)))
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{coin}}; \code{\link{death}},
  \code{\link{died}}; \code{\link{dice}};
  \code{\link{dummy}}; \code{\link{gpa}},
  \code{\link{grade}}, \code{\link{grade_letter}};
  \code{\link{group}}; \code{\link{height}},
  \code{\link{height_cm}}, \code{\link{height_in}};
  \code{\link{iq}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{normal}},
  \code{\link{normal_round}}; \code{\link{race}};
  \code{\link{sex}}; \code{\link{string}};
  \code{\link{valid}}
}
\keyword{likert}
\keyword{responses}

