% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_heat.R
\name{plot.as_tibble}
\alias{plot.as_tibble}
\title{Plots a as_tibble Object}
\usage{
\method{plot}{as_tibble}(x, ...)
}
\arguments{
\item{x}{The as_tibble object.}

\item{\ldots}{Arguments passed to \code{\link[wakefield]{table_heat}}.}
}
\description{
Plots a as_tibble object.
}
