% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_heat.R
\name{plot.tbl_df}
\alias{plot.tbl_df}
\title{Plots a tbl_df Object}
\usage{
\method{plot}{tbl_df}(x, ...)
}
\arguments{
\item{x}{The tbl_df object.}

\item{\ldots}{Arguments passed to \code{\link[wakefield]{table_heat}}.}
}
\description{
Plots a tbl_df object.
}
