% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/political.R
\name{political}
\alias{political}
\title{Generate Random Vector of Political Parties}
\usage{
political(n, x = c("Constitution", "Democrat", "Green", "Libertarian",
  "Republican"), prob = c(0.577269133302094, 0.410800432748879,
  0.00491084954793489, 0.00372590303330866, 0.0032936813677832),
  name = "Political")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random vector of political party elements.
}
\description{
Generate a random vector of political parties.
}
\details{
The political parties and probabilities used match approximate U.S.
political make-up of registered voters (2014).  The default make up is:

\tabular{lrr}{
  \bold{ Party} \tab \bold{N} \tab \bold{Percent} \cr
  Democrat     \tab 43,140,758 \tab 57.73\% \cr
  Republican   \tab 30,700,138 \tab 41.08\% \cr
  Constitution \tab   367,000 \tab   .49\% \cr
  Libertarian  \tab   278,446 \tab   .37\% \cr
  Green        \tab   246,145 \tab   .33\% \cr
}
}
\examples{
political(10)
barplot(table(political(10000)))
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{animal}}, \code{\link{pet}};
  \code{\link{answer}}; \code{\link{area}};
  \code{\link{car}}; \code{\link{coin}};
  \code{\link{color}}, \code{\link{primary}};
  \code{\link{death}}, \code{\link{died}};
  \code{\link{dice}}; \code{\link{dna}};
  \code{\link{dummy}}; \code{\link{education}};
  \code{\link{gpa}}, \code{\link{grade}},
  \code{\link{grade_letter}}; \code{\link{grade_level}};
  \code{\link{group}}; \code{\link{height}},
  \code{\link{height_cm}}, \code{\link{height_in}};
  \code{\link{income}}; \code{\link{iq}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{lower}},
  \code{\link{lower_factor}}, \code{\link{upper}},
  \code{\link{upper_factor}}; \code{\link{marital}};
  \code{\link{month}}; \code{\link{name}};
  \code{\link{normal}}, \code{\link{normal_round}};
  \code{\link{race}}; \code{\link{sat}};
  \code{\link{sentence}}; \code{\link{sex}};
  \code{\link{smokes}}; \code{\link{speed}},
  \code{\link{speed_kph}}, \code{\link{speed_mph}};
  \code{\link{state}}; \code{\link{string}};
  \code{\link{valid}}; \code{\link{zip_code}}
}
\keyword{democrat}
\keyword{poitical}
\keyword{politics}
\keyword{republican}

