% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{presidential_debates_2012}
\alias{presidential_debates_2012}
\title{2012 U.S. Presidential Debate Dialogue}
\format{
A character vector with 2911 elements
}
\usage{
data(presidential_debates_2012)
}
\description{
A dataset containing 2911 ordered sentences used by speakers during the three
2012 presidential debates.
}
\keyword{datasets}
