% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{print.available}
\alias{print.available}
\title{Prints an available Object.}
\usage{
\method{print}{available}(x, ...)
}
\arguments{
\item{x}{The available object}

\item{\ldots}{ignored}
}
\description{
Prints an available object.
}
