% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varname.R
\name{print.variable}
\alias{print.variable}
\title{Prints a variable Object}
\usage{
\method{print}{variable}(x, ...)
}
\arguments{
\item{x}{The \code{variable} object.}

\item{\ldots}{Ignored.}
}
\description{
Prints a \code{variable} object
}
