% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r_data.R
\name{r_data}
\alias{r_data}
\title{Pre-Seleceted Column Data Set}
\usage{
r_data(n = 500, ...)
}
\arguments{
\item{n}{The length to pass to the randomly generated vectors (number of rows).}

\item{\ldots}{A set of optionally named arguments.  Using \pkg{wakefield}
variable functions require no name or call parenthesis.}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}}.
}
\description{
Generate a data set with pre-set columns selected.
}
\details{
The pre-selected columns include:

\itemize{
  \item ID
  \item Race
  \item Age
  \item Sex
  \item Hour
  \item IQ
  \item Height
  \item Died
}

The user may use \ldots to add additional columns.  \code{r_data} is a
convenience function to quickly produce a data set.  For more specific usage
use the more flexible \code{\link[wakefield]{r_data_frame}} function.
}
\examples{
r_data()
r_data(10)
r_data(10, paragraph, Attending = valid)
}
\seealso{
\code{\link[wakefield]{r_data_frame}}
}

