% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_data_frame.R
\name{r_data_frame}
\alias{r_data_frame}
\title{Data Frame Production (From Variable Functions)}
\usage{
r_data_frame(n, ...)
}
\arguments{
\item{n}{The length to pass to the randomly generated vectors.}

\item{\ldots}{A set of optionally named arguments.  Using \pkg{wakefield}
variable functions require no name or call parenthesis.}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}}.
}
\description{
Produce a \code{\link[dplyr]{tbl_df}} data frame that allows the user to
lazily pass unnamed \pkg{wakefield} variable functions (optionally, without
call parenthesis).
}
\examples{
r_data_frame(
    n = 30,
    id,
    race,
    age,
    sex,
    hour,
    iq,
    height,
    died,
    Scoring = rnorm,
    Smoker = valid
)

r_data_frame(
    n = 30,
    id,
    race,
    age(x = 8:14),
    Gender = sex,
    Time = hour,
    iq,
    height(mean=50, sd = 10),
    died,
    Scoring = rnorm,
    Smoker = valid
)
}
\author{
Josh O'Brien and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/29617983/1000343}
}
\seealso{
\code{\link[wakefield]{r_list}}
}

