% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r_list.R
\name{r_list}
\alias{r_list}
\title{List Production (From Variable Functions)}
\usage{
r_list(n, ...)
}
\arguments{
\item{n}{The length to pass to the randomly generated vectors.}

\item{\ldots}{A set of optionally named arguments.  Using \pkg{wakefield}
variable functions require no name or call parenthesis.}
}
\value{
Returns a named list of equal length vectors.
}
\description{
Produce a named \code{\link[base]{list}} that allows the user to lazily pass
unnamed \pkg{wakefield} variable functions (optionally, without call
parenthesis).
}
\examples{
r_list(
    n = 30,
    d = rnorm,
    race,
    e = rnorm,
    p = rep("D", 30),
    rnorm
)
}
\author{
Josh O'Brien and Tyler Rinker <tyler.rinker@gmail.com>.
}
\references{
\url{http://stackoverflow.com/a/29617983/1000343}
}
\seealso{
\code{\link[wakefield]{r_data_frame}}
}
\keyword{list}

