% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sample_factor.R
\name{r_sample_factor}
\alias{r_sample_factor}
\title{Generate Random Factor Vector}
\usage{
r_sample_factor(n, x = LETTERS, prob = NULL, name = "Factor")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random actor vector of elements.
}
\description{
Generate a random vector and coerces to a factor.
}
\examples{
r_sample_factor(100, name = "Var")
table(r_sample_factor(x = c("Dog", "Cat", "Fish", "Bird"), n=1000))
r_sample_factor(x = c("B", "W"), prob = c(.7, .3), n = 25)
}
\seealso{
\code{\link[base]{sample}}
}
