% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_sample_replace.R
\name{r_sample_replace}
\alias{r_sample_replace}
\title{Generate Random Vector (Without Replacement)}
\usage{
r_sample_replace(n, x = 1:100, prob = NULL, replace = FALSE, name = "Sample")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{replace}{logical.  If \code{TRUE} sampling is done with replacement.
Default is without replacement.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random vector of elements.
}
\description{
Generate a random vector without replacement.
}
\examples{
r_sample(100, name = "Var")
table(r_sample(x = c("Dog", "Cat", "Fish", "Bird"), n=1000))
r_sample(x = c("B", "W"), prob = c(.7, .3), n = 25, name = "Race")
r_sample(25, x = c(TRUE, FALSE))
}
\seealso{
\code{\link[base]{sample}}
}
