% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/r_series.R
\name{r_series}
\alias{r_series}
\title{Data Frame Series (Repeated Measures)}
\usage{
r_series(fun, j, n, ..., rep.sep = "_")
}
\arguments{
\item{fun}{A \pkg{waefield} variable function.}

\item{j}{The number of columns to produce.}

\item{n}{The number of rows to produce.}

\item{rep.sep}{A separator to use for repeated variable names.  For example
if the \code{\link[wakefield]{age}} is used three times
(\code{r_data_frame(age, age, age)}), the name "Age" will be assigned to all
three columns.  The resuts in column names \code{c("Age_1", "Age_2", "Age_3")}.}

\item{\ldots}{Additional arguments passed to \code{fun}.}
}
\value{
Returns a \code{\link[dplyr]{tbl_df}}.
}
\description{
Produce a \code{\link[dplyr]{tbl_df}} data frame of repeated measures from a
wakefield variable function.
}
\examples{
r_series(grade, 5, 10)

## Custom name prefix
r_series(likert, 5, 10, name = "Question")

## Convert factors to integers
as_integer(r_series(likert_7, 5, 10))
}
\references{
\url{https://github.com/trinker/wakefield/issues/1/#issuecomment-96166910}
}
\seealso{
\code{\link[wakefield]{r_list}},
\code{\link[wakefield]{r_data_frame}}
\code{\link[wakefield]{r_dummy}}
}

