% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesname.R
\name{seriesname}
\alias{seriesname}
\title{Add Internal Name to Data Frame}
\usage{
seriesname(x, name)
}
\arguments{
\item{x}{A \code{\link[base]{data.frame}} to add a \code{seriesname}
attribute (i.e., \code{attributes(x)[["seriesname"]]})}

\item{name}{A name to assign to \code{attributes(x)[["seriesname"]]}.}
}
\value{
Returns a \code{\link[base]{data.frame}} with a
\code{attributes(x)[["seriesname"]]} assigned.
}
\description{
Adds \code{attributes(x)[["seriesname"]]} attribute to a
\code{\link[base]{data.frame}}.
}
\examples{
seriesname(mtcars, "Cars")
attributes(seriesname(mtcars, "Cars"))
}
