% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speed.R
\name{speed}
\alias{speed}
\alias{speed_kph}
\alias{speed_mph}
\title{Generate Random Vector of Speeds}
\usage{
speed(n, mean = 55, sd = 10, min = 0, max = NULL, digits = 0,
  name = "Speed")

speed_mph(n, mean = 55, sd = 10, min = 0, max = NULL, digits = 1,
  name = "Speed(mph)")

speed_kph(n, mean = 88.5, sd = 16, min = 0, max = NULL, digits = 1,
  name = "Speed(kph)")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{mean}{The mean value for the normal distribution to be drawn from.}

\item{sd}{The standard deviation of the normal distribution to draw from.}

\item{min}{A numeric lower boundary cutoff.  Resuts less than this value will be
replaced with \code{min}.}

\item{max}{A numeric upper boundary cutoff.  Resuts greater than this value will
be replaced with \code{max}.}

\item{digits}{Integer indicating the number of decimal places to be used.
Negative values are allowed (see \code{\link[base]{round}}).}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random normal vector of speed elements.
}
\description{
\code{speed} and \code{speed_in} - Generate a random normal vector of
speeds in inches.

\code{speed_cm} - Generate a random normal vector of speeds in centimeters.
}
\note{
\code{speed} rounds to nearest whoe number.  \code{speed_in} &
\code{speed_in} round to the nearest tenths.
}
\examples{
speed(10)
hist(speed(10000))
interval(speed, 5, n = 1000)
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{animal}}, \code{\link{pet}};
  \code{\link{answer}}; \code{\link{area}};
  \code{\link{car}}; \code{\link{coin}};
  \code{\link{color}}, \code{\link{primary}};
  \code{\link{death}}, \code{\link{died}};
  \code{\link{dice}}; \code{\link{dna}};
  \code{\link{dummy}}; \code{\link{education}};
  \code{\link{gpa}}, \code{\link{grade}},
  \code{\link{grade_letter}}; \code{\link{grade_level}};
  \code{\link{group}}; \code{\link{height}},
  \code{\link{height_cm}}, \code{\link{height_in}};
  \code{\link{income}}; \code{\link{iq}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{lower}},
  \code{\link{lower_factor}}, \code{\link{upper}},
  \code{\link{upper_factor}}; \code{\link{marital}};
  \code{\link{month}}; \code{\link{name}};
  \code{\link{normal}}, \code{\link{normal_round}};
  \code{\link{political}}; \code{\link{race}};
  \code{\link{sat}}; \code{\link{sentence}};
  \code{\link{sex}}; \code{\link{smokes}};
  \code{\link{state}}; \code{\link{string}};
  \code{\link{valid}}; \code{\link{zip_code}}
}
\keyword{speed}

