% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{state_populations}
\alias{state_populations}
\title{State Populations (2010)}
\format{
A data frame with 50 rows and 3 variables
}
\usage{
data(state_populations)
}
\description{
A dataset containing U.S. state populations.
}
\details{
\itemize{
  \item State. The 50 U.S. states.
  \item Population. Population of state.
  \item Proportion. Proportion of total U.S. population.
}
}
\references{
https://en.wikipedia.org/wiki/List_of_U.S._states_and_territories_by_population
}
\keyword{datasets}
