% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/string.R
\name{string}
\alias{string}
\title{Generate Random Vector of Strings}
\usage{
string(n, x = "[A-Za-z0-9]", length = 10, name = "String")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A character vector specifying character classes to draw elements
from.}

\item{length}{Integer vector, desired string lengths.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random character vector of string elements.
}
\description{
Generate a random vector of strings.
}
\examples{
string(10)
}
\seealso{
\code{\link[stringi]{stri_rand_strings}}

Other variable.functions: \code{\link{dummy}};
  \code{\link{lorem_ipsum}}, \code{\link{paragraph}};
  \code{\link{race}}; \code{\link{sex}};
  \code{\link{valid}}
}
\keyword{character}
\keyword{string}

