% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/valid.R
\name{valid}
\alias{valid}
\title{Generate Random Logical Vector}
\usage{
valid(n, prob = NULL, name = "Valid")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random logical vector of elements.
}
\description{
Generate a random logical (\code{TRUE}/\code{FALSE}).
}
\examples{
valid(10)
100*table(valid(n <- 1000))/n
}
\seealso{
Other variable.functions: \code{\link{dummy}};
  \code{\link{lorem_ipsum}}, \code{\link{paragraph}};
  \code{\link{race}}; \code{\link{sex}};
  \code{\link{string}}
}
\keyword{logical}
\keyword{valid}

