% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varname.R
\name{varname}
\alias{varname}
\title{Title}
\usage{
varname(x, name)
}
\arguments{
\item{x}{A vector to add a code{varname} attribute (i.e.,
\code{attributes(x)[["varname"]]})}

\item{name}{A name to assign to \code{attributes(x)[["varname"]]}.}
}
\value{
Returns a vector with a \code{attributes(x)[["varname"]]} assigned.
}
\description{
Description
}
\examples{
varname(1:10, "A")
varname(LETTERS, "Caps")
}
\seealso{
Other res: \code{\link{race}}
}

