% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{package}
\name{wakefield-package}
\alias{wakefield}
\alias{wakefield-package}
\title{wakefield: Generate Random Data Sets}
\description{
Generates random data sets including: data.frames, lists, and vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/trinker/wakefield}
  \item Report bugs at \url{https://github.com/trinker/wakefield/issues}
}

}
\author{
\strong{Maintainer}: Tyler Rinker \email{tyler.rinker@gmail.com}

Other contributors:
\itemize{
  \item Josh O'Brien [contributor]
  \item Ananda Mahto [contributor]
  \item Matthew Sigal [contributor]
  \item Jonathan Carroll [contributor]
  \item Scott Westenberger [contributor]
}

}
\keyword{internal}
