% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{package}
\name{wakefield}
\alias{package-wakefield}
\alias{wakefield}
\alias{wakefield-package}
\title{Generate Random Data Sets}
\description{
Generates random data sets including: data.frames, lists, and vectors.
}

