% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Generate Random Dummy Coded Vector}
\usage{
dummy(n, prob = NULL, name = "Dummy")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random dummy vector of (0/1) elements.
}
\description{
Generate a random dummy coded (0/1) vector.
}
\examples{
dummy(100, name = "Var")
table(dummy(1000))
}
\seealso{
\code{\link[base]{sample.int}}

Other variable.functions: \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{race}};
  \code{\link{sex}}; \code{\link{string}};
  \code{\link{valid}}
}

