% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{grady_augmented}
\alias{grady_augmented}
\title{Augmented List of Grady Ward's English Words and Mark Kantrowitz's Names List}
\format{
A character vector with 122806 elements
}
\usage{
data(grady_augmented)
}
\description{
A dataset containing a vector of Grady Ward's English words augmented with
\pkg{qdapDictionaries}'s  \code{DICTIONARY}, Mark Kantrowitz's names list,
other proper nouns, and contractions.
}
\details{
A dataset containing a vector of Grady Ward's English words
augmented with proper nouns (U.S. States, Countries, Mark Kantrowitz's Names
List, and months) and contractions. That dataset is augmented to increase the
data set size.
}
\references{
Moby Thesaurus List by Grady Ward https://www.gutenberg.org \cr \cr
List of names from Mark Kantrowitz http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/.
A copy of the http://www.cs.cmu.edu/afs/cs/project/ai-repository/ai/areas/nlp/corpora/names/readme.txt
per the author's request.
}
\keyword{datasets}
