% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wakefield-package.R
\docType{data}
\name{name_neutral}
\alias{name_neutral}
\title{Gender Neutral Names}
\format{
A character vector with 662 elements
}
\usage{
data(name_neutral)
}
\description{
A dataset containing a character vector gender neutral names according to the
U.S. Census.
}
\references{
http://www.census.gov
}
\keyword{datasets}
