% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/race.R
\name{race}
\alias{race}
\title{Generate Random Vector of Races}
\usage{
race(name = "Race", races = NULL, prob = NULL, n)
}
\arguments{
\item{name}{The name to assign to the output vector's \code{varname}
attribute (i.e., \code{attributes(x)[["varname"]]}).  This is auto assigned as
the column/vector name when used inside of \code{r_data_frame} or \code{r_list}.}

\item{races}{A vector of races to chose from.  See \bold{details} for the
default races.}

\item{prob}{A vector of probabilities to chose from.  See \bold{details} for
the default probabilities.}

\item{n}{The number elements to generate.  This can be auto set if used
inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random vector of elements.
}
\description{
Generate a random vector of races
}
\details{
The races and probabilities used match approximate U.S. racial
make-up.  The default make up is:

\tabular{lr}{
  \bold{Race}     \tab \bold{Percent}\cr
  White     \tab 63.70 \%\cr
  Hispanic  \tab 16.30 \%\cr
  Black     \tab 12.20 \%\cr
  Asian     \tab 4.70 \%\cr
  Bi-Racial \tab 1.90 \%\cr
  Other     \tab .20 \%\cr
  Native    \tab .70 \%\cr
  Hawaiian  \tab .15 \%
}
}
\examples{
race("Race_Column", n=100)
table(race("Race_Column", n=1000))
race(races = c("B", "W"), prob = c(.7, .3), n = 25)
}
\seealso{
Other res: \code{\link{varname}}
}
\keyword{race}

