% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/race.R
\name{race}
\alias{race}
\title{Generate Random Vector of Races}
\usage{
race(n, x = c("White", "Hispanic", "Black", "Asian", "Bi-Racial", "Native",
  "Other", "Hawaiian"), prob = c(0.637, 0.163, 0.122, 0.047, 0.019, 0.007,
  0.002, 0.0015), name = "Race")
}
\arguments{
\item{n}{The number elements to generate.  This can be globally set within
the environment of \code{r_data_frame} or \code{r_list}.}

\item{x}{A vector of elements to chose from.}

\item{prob}{A vector of probabilities to chose from.}

\item{name}{The name to assign to the output vector's \code{varname}
attribute.  This is used to auto assign names to the column/vector name when
used inside of \code{r_data_frame} or \code{r_list}.}
}
\value{
Returns a random factor vector of elements.
}
\description{
Generate a random vector of races.
}
\details{
The races and probabilities used match approximate U.S. racial
make-up.  The default make up is:

\tabular{lr}{
  \bold{Race}     \tab \bold{Percent}\cr
  White     \tab 63.70 \%\cr
  Hispanic  \tab 16.30 \%\cr
  Black     \tab 12.20 \%\cr
  Asian     \tab 4.70 \%\cr
  Bi-Racial \tab 1.90 \%\cr
  Native    \tab .70 \%\cr
  Other     \tab .20 \%\cr
  Hawaiian  \tab .15 \%\cr
}
}
\examples{
race(10)
100*table(race(n <- 10000))/n
}
\seealso{
Other variable.functions: \code{\link{age}};
  \code{\link{animal}}, \code{\link{pet}};
  \code{\link{answer}}; \code{\link{area}};
  \code{\link{car}}; \code{\link{children}};
  \code{\link{coin}}; \code{\link{color}},
  \code{\link{primary}}; \code{\link{death}},
  \code{\link{died}}; \code{\link{dice}};
  \code{\link{dna}}; \code{\link{dummy}};
  \code{\link{education}}; \code{\link{eye}};
  \code{\link{gpa}}, \code{\link{grade}},
  \code{\link{grade_letter}}; \code{\link{grade_level}};
  \code{\link{group}}; \code{\link{hair}};
  \code{\link{height}}, \code{\link{height_cm}},
  \code{\link{height_in}}; \code{\link{income}};
  \code{\link{iq}}; \code{\link{level}};
  \code{\link{likert}}, \code{\link{likert_5}},
  \code{\link{likert_7}}; \code{\link{lorem_ipsum}},
  \code{\link{paragraph}}; \code{\link{lower}},
  \code{\link{lower_factor}}, \code{\link{upper}},
  \code{\link{upper_factor}}; \code{\link{marital}};
  \code{\link{month}}; \code{\link{name}};
  \code{\link{normal}}, \code{\link{normal_round}};
  \code{\link{political}}; \code{\link{sat}};
  \code{\link{sentence}}; \code{\link{sex}};
  \code{\link{smokes}}; \code{\link{speed}},
  \code{\link{speed_kph}}, \code{\link{speed_mph}};
  \code{\link{state}}; \code{\link{string}};
  \code{\link{valid}}; \code{\link{year}};
  \code{\link{zip_code}}
}
\keyword{race}

